requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.3.29
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.3
datasets==2.16.1
Deprecated==1.2.14
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.109.2
filelock==3.13.1
FLAML==2.1.1
flatbuffers==23.5.26
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2023.10.0
git-filter-repo==2.38.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.28.1
googleapis-common-protos==1.62.0
graphlib-backport==1.0.3
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
langchain==0.1.9
langchain-community==0.0.22
langchain-core==0.1.26
langchain-openai==0.0.3
langchainhub==0.1.14
langsmith==0.1.5
lark==1.1.9
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
multitasking==0.0.11
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.24.4
oauthlib==3.2.2
onnxruntime==1.17.0
openai==1.12.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.14
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pathlib==1.0.1
peewee==3.17.0
pillow==10.2.0
pinecone-client==3.0.1
pkgutil_resolve_name==1.3.10
posthog==3.4.2
protobuf==4.25.2
pulsar-client==3.4.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyautogen==0.2.14
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pypdf==4.0.2
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.36.3
streamlit==1.30.0
style==1.1.0
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
textwrap3==0.9.2
tiktoken==0.5.2
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
typer==0.9.0
types-requests==2.31.0.20240218
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
update==0.0.1
urllib3==2.1.0
uvicorn==0.27.1
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
